Periodic payment status change check should be running even if user refreshes the page while the payment is panding.
So far it was triggered by JS upon transaction hash submission to the BE, now it runs on page load also when needed.
