+++ pretix_eth/templates/pretix_eth/pending.html
@@ -9,8 +9,9 @@
 <a id="pretix-order-detail-url" data-order-detail-url="{% url 'api-v1:order-detail' organizer=payment.order.event.organizer.slug event=payment.order.event.slug code=payment.order.code %}"></a>
 
   {% if submitted_transation_hash %}
+<script type="module" src="/static/pretix_eth/web3modal/run_periodic_check.js"></script>
 <div>
-  <p>A payment for this order has already been submitted in transaction <strong>{{ submitted_transation_hash }}</strong>.</p>
+  <p>A payment for this order has already been submitted in transaction <strong><span id="pretix-eth-submitted-transaction-hash">{{ submitted_transation_hash }}</span></strong>.</p>
 </div>
   {% endif %}
